requirements.txt
alembic==1.11.1
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.9.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Share==0.1.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
Pillow==9.5.0
PyJWT==2.7.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
SQLAlchemy==2.0.15
typing_extensions==4.6.2
Werkzeug==2.3.4
